+++ backend/src/main/java/com/yoga/backend/mypage/livelectures/dto/LiveLectureCreateDto.java
@@ -17,23 +17,21 @@ public class LiveLectureCreateDto {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int liveId; // 화상 강의 ID
+    private Long liveId; // 화상 강의 ID
 
     private String liveTitle; // 강의 제목
-
     private String liveContent; // 강의 내용
 
-    private Long startDate; // 시작 날짜
-    private Long endDate; // 종료 날짜
+    private Long startDate; // 시작 날짜 (밀리초 단위의 timestamp)
+    private Long endDate; // 종료 날짜 (밀리초 단위의 timestamp)
 
-    private Long startTime; // 강의 시작 시간
-    private Long endTime; // 강의 종료 시간
+    private Long startTime; // 강의 시작 시간 (밀리초 단위의 timestamp)
+    private Long endTime; // 강의 종료 시간 (밀리초 단위의 timestamp)
 
     private Integer maxLiveNum; // 최대 수강자 수
-    private Long regDate; // 강의 등록 시간
+    private Long regDate; // 강의 등록 시간 (밀리초 단위의 timestamp)
 
-    private Integer userId; // 강사 ID
+    private int userId; // 강사 ID (Integer에서 Long으로 변경)
 
     private String availableDay; // 가능한 강의 요일
-
 }
